format visited domain list in getter as well as get default state
